feat: Users can edit categories

+++ server/controllers/category.controller..js
@@ -1,9 +1,10 @@
 const jwt = require("jsonwebtoken");
 const { pool } = require("../db");
+const slugify = require("slugify");
 
-exports.getTransactionsByCategory = async (req, res) => {
+exports.getCategoryBySlug = async (req, res) => {
 	const token = req.headers["authorization"].split(" ")[1];
-	const { category } = req.params;
+	const { category_slug } = req.params;
 
 	jwt.verify(token, process.env.JWT_SECRET, async (err, decoded) => {
 		if (err) {
@@ -15,25 +16,61 @@ exports.getTransactionsByCategory = async (req, res) => {
 
 		// The token is valid, so we can now query the database
 		try {
-			const transactions = await pool.query(
+			const category = await pool.query(
 				`
-				SELECT transaction.id, title, note, amount, created_at, updated_at, type,
-				category.name AS category_name, category.slug AS category_slug
-				FROM transaction, category
-				WHERE category.slug=$1
-				AND transaction.category_id = category.id
-				AND transaction.user_id = $2
-				ORDER BY created_at
-				DESC LIMIT 10
+				SELECT *
+				FROM category
+				WHERE category.slug = $1
+				AND category.user_id = $2
 				`,
-				[category, decoded.id]
+				[category_slug, decoded.id]
 			);
 
-			if (!transactions.rowCount) {
+			if (!category.rowCount) {
 				return res.json({ ok: false, message: "Transaction not found" });
 			}
 
-			res.status(200).json({ ok: true, transactions: transactions.rows });
+			res.status(200).json({ ok: true, category: category.rows[0] });
+		} catch (err) {
+			console.log(err);
+			res.status(500).json({ ok: false, message: "Error" });
+		}
+	});
+};
+
+exports.updateCategory = async (req, res) => {
+	const token = req.headers["authorization"].split(" ")[1];
+	const { category_slug } = req.params;
+	const { name } = req.body;
+	const slug = slugify(name, { lower: true });
+
+	jwt.verify(token, process.env.JWT_SECRET, async (err, decoded) => {
+		if (err) {
+			return res.json({
+				ok: false,
+				message: "Invalid token",
+			});
+		}
+
+		// The token is valid, so we can update the category
+		try {
+			const updatedCategory = await pool.query(
+				`
+				UPDATE category
+				SET name = $1,
+					slug = $2
+				WHERE category.slug = $3
+				AND category.user_id = $4
+				RETURNING *
+				`,
+				[name, slug, category_slug, decoded.id]
+			);
+
+			if (!updatedCategory.rowCount) {
+				return res.json({ ok: false, message: "Category not found" });
+			}
+
+			res.status(200).json({ ok: true, category: updatedCategory.rows[0] });
 		} catch (err) {
 			console.log(err);
 			res.status(500).json({ ok: false, message: "Error" });
@@ -43,7 +80,7 @@ exports.getTransactionsByCategory = async (req, res) => {
 
 exports.deleteCategory = async (req, res) => {
 	const token = req.headers["authorization"].split(" ")[1];
-	const { category } = req.params;
+	const { category_slug } = req.params;
 
 	jwt.verify(token, process.env.JWT_SECRET, async (err, decoded) => {
 		if (err) {
@@ -55,6 +92,26 @@ exports.deleteCategory = async (req, res) => {
 
 		// The token is valid, so we can now delete the category from the database
 		try {
+			// If the category has transactions, we can't delete it
+			const transactions = await pool.query(
+				`
+				SELECT category.* 
+				FROM category
+				LEFT JOIN transaction
+				ON category.id = transaction.category_id
+				WHERE category.slug = $1
+				AND transaction.user_id = $2
+				`,
+				[category_slug, decoded.id]
+			);
+
+			if (transactions.rowCount) {
+				return res.json({
+					ok: false,
+					message: "You can't delete a category that has transactions",
+				});
+			}
+
 			const deletedCategory = await pool.query(
 				`
 				DELETE FROM category
@@ -62,7 +119,7 @@ exports.deleteCategory = async (req, res) => {
 				AND category.user_id = $2
 				RETURNING category.slug
 				`,
-				[category, decoded.id]
+				[category_slug, decoded.id]
 			);
 
 			if (!deletedCategory.rowCount) {
